+++ techgrains/lib/com/techgrains/service/tg_service.dart
@@ -12,6 +12,7 @@ import 'package:techgrains/com/techgrains/service/request/tg_get_request.dart';
 import 'package:techgrains/com/techgrains/service/request/tg_post_request.dart';
 import 'package:techgrains/com/techgrains/service/request/tg_put_request.dart';
 import 'package:techgrains/com/techgrains/service/request/tg_request.dart';
+import 'package:techgrains/com/techgrains/service/request/tg_upload_file_request.dart';
 import 'package:techgrains/com/techgrains/service/request/tg_upload_request.dart';
 import 'package:techgrains/com/techgrains/service/response/tg_response.dart';
 
@@ -24,8 +25,7 @@ class TGService<T extends TGResponse, E extends TGError> {
 
   TGService(this.creatorT, this.creatorE);
 
-  Client _getClient(String url, String method) =>
-      TGClientFactory.getClient(url, method);
+  Client _getClient(String url, String method) => TGClientFactory.getClient(url, method);
 
   static init(
       {required String baseUrl,
@@ -39,24 +39,20 @@ class TGService<T extends TGResponse, E extends TGError> {
 
     // For Mock
     TGMockService.applyMock = applyMock;
-    if (mockMappingsFile != null)
-      TGMockService.loadMockMappings(mockMappingsFile);
+    if (mockMappingsFile != null) TGMockService.loadMockMappings(mockMappingsFile);
 
     // For HTTP Client
     TGHttpClient.badCertificateCallbackEnabled = badCertificateCallbackEnabled;
   }
 
-  Future<T> get(
-      {required TGGetRequest request, onSuccess(T)?, onError(T)?}) async {
+  Future<T> get({required TGGetRequest request, onSuccess(T)?, onError(T)?}) async {
     Uri uri = Uri.parse(request.getUrl());
     TGLog.t("GET", uri);
-    final httpRes = await _getClient(request.getUri(), "GET")
-        .get(uri, headers: request.headers());
+    final httpRes = await _getClient(request.getUri(), "GET").get(uri, headers: request.headers());
     return _performCallback(httpRes, onError, onSuccess);
   }
 
-  Future<T> post(
-      {required TGPostRequest request, onSuccess(T)?, onError(T)?}) async {
+  Future<T> post({required TGPostRequest request, onSuccess(T)?, onError(T)?}) async {
     Uri uri = Uri.parse(request.getUrl());
     TGLog.t("POST", uri);
     final httpRes = await _getClient(request.getUri(), "POST").post(
@@ -67,8 +63,7 @@ class TGService<T extends TGResponse, E extends TGError> {
     return _performCallback(httpRes, onError, onSuccess);
   }
 
-  Future<T> put(
-      {required TGPutRequest request, onSuccess(T)?, onError(T)?}) async {
+  Future<T> put({required TGPutRequest request, onSuccess(T)?, onError(T)?}) async {
     Uri uri = Uri.parse(request.getUrl());
     TGLog.t("PUT", uri);
     final httpRes = await _getClient(request.getUri(), "PUT").put(
@@ -79,35 +74,40 @@ class TGService<T extends TGResponse, E extends TGError> {
     return _performCallback(httpRes, onError, onSuccess);
   }
 
-  Future<T> delete(
-      {required TGDeleteRequest request, onSuccess(T)?, onError(T)?}) async {
+  Future<T> delete({required TGDeleteRequest request, onSuccess(T)?, onError(T)?}) async {
     Uri uri = Uri.parse(request.getUrl());
     TGLog.t("DELETE", uri);
-    final httpRes = await _getClient(request.getUri(), "DELETE")
-        .delete(uri, headers: request.headers());
+    final httpRes = await _getClient(request.getUri(), "DELETE").delete(
+      uri,
+      body: request.body(),
+      headers: request.headers(),
+    );
     return _performCallback(httpRes, onError, onSuccess);
   }
 
-  Future<T> upload(
-      {required TGUploadRequest request, onSuccess(T)?, onError(T)?}) async {
-    var multipartRequest = http.MultipartRequest(
-        "POST",
-        Uri.parse(
-            TGRequest.prepareUrl(TGRequest.defaultBaseUrl, request.getUri())));
+  Future<T> upload({required TGUploadRequest request, onSuccess(T)?, onError(T)?}) async {
+    var multipartRequest = http.MultipartRequest("POST", Uri.parse(TGRequest.prepareUrl(TGRequest.defaultBaseUrl, request.getUri())));
     multipartRequest.files.add(request.file());
     StreamedResponse httpRes = await multipartRequest.send();
     return _performCallbackForStreamedResponse(httpRes, onError, onSuccess);
   }
 
-  T _performCallback(
-      Response httpRes, onError(dynamic T)?, onSuccess(dynamic T)?) {
+  Future<T> uploadFile({required TGUploadFileRequest request, onSuccess(T)?, onError(E)?}) async {
+    var multipartRequest = http.MultipartRequest("POST", Uri.parse(TGRequest.prepareUrl(TGRequest.defaultBaseUrl, request.getUri())));
+    multipartRequest.headers.addAll(request.headers()!);
+    multipartRequest.fields.addAll(request.body());
+    multipartRequest.files.add(request.file());
+    StreamedResponse httpRes = await multipartRequest.send();
+    return _performCallbackForUploadFileResponse(httpRes, onError, onSuccess);
+  }
+
+  T _performCallback(Response httpRes, onError(dynamic T)?, onSuccess(dynamic T)?) {
     T t = _prepareResponse(httpRes);
     t.hasError ? onError!(t) : onSuccess!(t);
     return t;
   }
 
-  T _performCallbackForStreamedResponse(
-      StreamedResponse httpRes, onError(dynamic T)?, onSuccess(dynamic T)?) {
+  T _performCallbackForStreamedResponse(StreamedResponse httpRes, onError(dynamic T)?, onSuccess(dynamic T)?) {
     T t = creatorT();
     try {
       _populateResponse(t, httpRes);
@@ -118,6 +118,29 @@ class TGService<T extends TGResponse, E extends TGError> {
       // Ignores error if not able to populate response, validate response or decode json. Populated response must be returned.
     }
     t.hasError ? onError!(t) : onSuccess!(t);
+    return t;
+  }
+
+  Future<T> _performCallbackForUploadFileResponse(StreamedResponse httpRes, onError(dynamic E)?, onSuccess(dynamic T)?) async {
+    T t = creatorT();
+    try {
+      _populateResponse(t, httpRes);
+      await httpRes.stream.transform(utf8.decoder).listen((value) {
+        t.body = value;
+      });
+      _validateResponse(t);
+    } catch (e) {
+      TGLog.e(e);
+      // Ignores error if not able to populate response, validate response or decode json. Populated response must be returned.
+    }
+    if (t.hasError) {
+      E e = creatorE();
+      final timestamp = new DateTime.now().millisecondsSinceEpoch;
+      final errorResponse = {"httpStatus": httpRes.statusCode, "timestamp": timestamp, "message": t.body!};
+      onError!(e.fromJson(errorResponse));
+    } else {
+      onSuccess!(t);
+    }
     return t;
   }
 
@@ -151,10 +174,8 @@ class TGService<T extends TGResponse, E extends TGError> {
           t.fromJson(jsonDecode(t.body!));
         }
       } catch (e) {
-        TGLog.e("Unable to prepare - " +
-            (t.hasError ? E.toString() : T.toString()));
-        TGLog.e("Please check 'fromJson(dynamic json)' implementation of - " +
-            (t.hasError ? E.toString() : T.toString()));
+        TGLog.e("Unable to prepare - " + (t.hasError ? E.toString() : T.toString()));
+        TGLog.e("Please check 'fromJson(dynamic json)' implementation of - " + (t.hasError ? E.toString() : T.toString()));
         TGLog.e(e);
         t.hasError = true;
         print(t.body);